reworks deterministic primes testing, experimental
This rework of the deterministic primes test, which is still a variant of the
classic, simple, naive primes test (not, for instance, any variant of MKS), but
this one stores primes more efficiently, giving the test a greater reach. A
problem is that the Bignum library in use manages its memory privately, and
automatically, which means that if we allocate too much memory for allocating
mpz_t pointers, the Bignum library has a greater chance of failing an allocation,
which will bring the processing to an end. What's a better way to manage this?

As it is, the largest prime this particular implementation was able to generate
as part of its testing feedstock, had 13 digits. This is a large-ish prime, for
a naive algorithm, running for only a couple hours, but it is entirely
inadequate for the task. This would allow primality tests of numbers with
around 26 digits.  Fibonacci numbers grow so fast that the usefulness of this
test is minimal.

I'm checking this in to save the work, because it's interesting, but the final
implementation should not include this more or less useless code.

This change also adds debugging symbols to the fizzbuzz binary.