adds Doxygen comments, removes deterministic primality checking support
The deterministic primality checks, as noted in the last checkin, aren't
terribly useful, because they're not effective. Meanwhile, according to
https://gmplib.org/manual/Number-Theoretic-Functions.html , the probability
that the method used to check primality will falsely report a composite as
prime is less than 1/(4^50). Odds that should be good enough for fizzbuzz.

Added Doxygen comments to the remaining functions, and Doxygenized some
existing comments.

Re-ordered function list so that the file-scoped 'progname' will be located
next to the other file-scoped variable 'opts', and now they're all easy to
find.