makes linefeed addition a non-default option
Upon considering the spec again, I realized that it says to print "Buzz" and
"Fizz", as opposed to "Buzz\n" and "Fizz\n", thus it seems I have inferred
something I should not have, and removed the linefeed from the default output.
I also added a new commandline option '-lf' which, when used, adds that
linefeed that had been the default before.

This new option has one other side effect. In the case where 3 and 5 are the
numbers being considerred, when both their divisibility and primality are
reported, I had placed a space between the divisibility report and the
primality report, a la: "Fizz BuzzFizz\n" for 5.  This has been corrected to
"FizzBuzzFizz" for 5. The -lf option, in addition to adding linefeeds, also
restores this space between these dually reported properties for 3 and 5.