adds option -a, which adds asterisk to some primes
Prime numbers larger than the smallest cannot be declared with *perfect*
reliability to be definitely prime. They are merely almost certainly prime. I
had been adding an asterisk to the BuzzFizz reporting, for larger primes, when
the verbose option is 1 or more. It seems reasonable, however, that the added
verbosity may be desired without necessarily forcing the asterisk to appear.
So, the asterisk no longer is added as a result of added verbosity and is
instead given its own option: '-a'. This option is off by default.